+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 0.3 (unreleased)
 ================
 
-- Add `add_aggregation` to `Dataset` class. By @huard
+- Add `add_aggregation` and `add_variable_agg` to `Dataset` class. By @huard
 - Add `add_scan` to `Dataset` class. By @huard