+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.3 (unreleased)
+================
+
+- Closing the dataset returned by `open_ncml` will close the underlying opened files. By @huard
+
+
 0.2 (2023-02-23)
 ================